backend/requirements.txt
Django>=5.2.8
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
redis>=5.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
